Fit per-window Dock changes inside the existing file-length budgets
No budget increases: the window-Dock browser resolvers, selector
conflict check, owner-anchored payload, and Dock browser close move to
a new TerminalController+WindowDockBrowserRouting.swift; the shared
focus/reveal owner-anchoring and the Dock close branch move into
TerminalController+ControlSurfaceDock.swift as reusable helpers; the
main-window ForTesting seams move to a dedicated
AppDelegate+MainWindowTestingSupport.swift per the debug-seam policy;
and the cross-window routing socket tests move to a new
WindowDockRoutingSocketTests.swift. Behavior is unchanged.
