+++ Sources/TerminalController.swift
@@ -3168,11 +3168,7 @@ class TerminalController {
             }
             // A window-Dock owner id IS its owning window's id, so a Dock-scoped
             // workspace_id routes to that window rather than the caller's.
-            if let tm = v2MainSync({
-                AppDelegate.shared?.existingWindowDock(forWindowId: wsId) != nil
-                    ? AppDelegate.shared?.tabManagerFor(windowId: wsId)
-                    : nil
-            }) {
+            if let tm = v2MainSync({ AppDelegate.shared?.tabManagerForWindowDockOwner(wsId) }) {
                 return tm
             }
         }
@@ -3223,8 +3219,7 @@ class TerminalController {
             }
             // A window-Dock owner id IS its owning window's id, so a Dock-scoped
             // workspace_id routes to that window rather than the caller's.
-            if AppDelegate.shared?.existingWindowDock(forWindowId: workspaceId) != nil,
-               let tm = AppDelegate.shared?.tabManagerFor(windowId: workspaceId) {
+            if let tm = AppDelegate.shared?.tabManagerForWindowDockOwner(workspaceId) {
                 return tm
             }
         }
@@ -5262,7 +5257,9 @@ class TerminalController {
 
     /// Value snapshot of a resolved browser surface for socket-worker handlers:
     /// resolution happens on the main actor, the JS-evaluating body runs off it.
-    private struct V2BrowserPanelContext {
+    // Internal (not private): the window-Dock browser resolvers live in
+    // TerminalController+WindowDockBrowserRouting.swift.
+    struct V2BrowserPanelContext {
         let workspaceId: UUID
         let surfaceId: UUID
         let browserPanel: BrowserPanel
@@ -5302,144 +5299,6 @@ class TerminalController {
         )
     }
 
-    private func v2ResolveWindowDockBrowserPanelContext(
-        params: [String: Any],
-        tabManager: TabManager
-    ) -> (handled: Bool, context: V2BrowserPanelContext?, error: V2CallResult?) {
-        let requestedWorkspaceID = v2UUID(params, "workspace_id")
-        let requestedSurfaceID = v2UUID(params, "surface_id") ?? v2UUID(params, "tab_id")
-        let requestedPaneID = v2UUID(params, "pane_id")
-
-        let dockBySurface = requestedSurfaceID.flatMap { windowDockContainingPanel($0) }
-        let dockByPane = requestedPaneID.flatMap { windowDockContainingPane($0) }
-        let routesWindowDock = requestedWorkspaceID.map(AppDelegate.isWindowDockRoutingId) == true
-            || dockBySurface != nil
-            || dockByPane != nil
-        guard routesWindowDock else {
-            return (false, nil, nil)
-        }
-
-        let dockByOwner = requestedWorkspaceID.flatMap { AppDelegate.shared?.existingWindowDock(forWindowId: $0) }
-        // Explicit selectors that name two different windows' Docks fail closed
-        // rather than silently acting on one of them. (The legacy alias pins no
-        // specific Dock, so it never conflicts.)
-        if windowDockSelectorsConflict(
-            requestedWorkspaceID: requestedWorkspaceID,
-            requestedWindowID: v2UUID(params, "window_id"),
-            dockByOwner: dockByOwner,
-            dockBySurface: dockBySurface,
-            dockByPane: dockByPane
-        ) {
-            return (true, nil, .err(code: "invalid_params", message: dockConflictingRoutingSelectorsMessage(), data: nil))
-        }
-        guard let dock = dockBySurface ?? dockByPane ?? dockByOwner
-                ?? AppDelegate.shared?.existingWindowDock(for: tabManager) else {
-            return (true, nil, .err(code: "not_found", message: "No focused browser surface", data: nil))
-        }
-
-        let surfaceId: UUID?
-        if let requestedSurfaceID {
-            surfaceId = requestedSurfaceID
-        } else if let requestedPaneID {
-            guard let pane = dock.bonsplitController.allPaneIds.first(where: { $0.id == requestedPaneID }) else {
-                return (true, nil, .err(code: "not_found", message: "Pane not found", data: ["pane_id": requestedPaneID.uuidString]))
-            }
-            guard let selectedTab = dock.bonsplitController.selectedTab(inPane: pane),
-                  let selectedSurface = dock.panel(for: selectedTab.id)?.id else {
-                return (true, nil, .err(code: "not_found", message: "Pane has no selected surface", data: ["pane_id": requestedPaneID.uuidString]))
-            }
-            surfaceId = selectedSurface
-        } else {
-            surfaceId = dock.focusedPanelId
-        }
-
-        guard let surfaceId else {
-            return (true, nil, .err(code: "not_found", message: "No focused browser surface", data: nil))
-        }
-        guard let browserPanel = dock.browserPanel(for: surfaceId) else {
-            return (true, nil, .err(code: "invalid_params", message: "Surface is not a browser", data: ["surface_id": surfaceId.uuidString]))
-        }
-        return (
-            true,
-            V2BrowserPanelContext(
-                workspaceId: dock.workspaceId,
-                surfaceId: surfaceId,
-                browserPanel: browserPanel,
-                webView: browserPanel.webView
-            ),
-            nil
-        )
-    }
-
-    private func v2ResolveWindowDockBrowserTabStore(
-        params: [String: Any],
-        tabManager: TabManager
-    ) -> (handled: Bool, dock: DockSplitStore?, error: V2CallResult?) {
-        let requestedWorkspaceID = v2UUID(params, "workspace_id")
-        let requestedSurfaceID = v2UUID(params, "surface_id")
-            ?? v2UUID(params, "tab_id")
-            ?? v2UUID(params, "target_surface_id")
-        let requestedPaneID = v2UUID(params, "pane_id")
-            ?? v2UUID(params, "target_pane_id")
-
-        let dockBySurface = requestedSurfaceID.flatMap { windowDockContainingPanel($0) }
-        let dockByPane = requestedPaneID.flatMap { windowDockContainingPane($0) }
-        let routesWindowDock = requestedWorkspaceID.map(AppDelegate.isWindowDockRoutingId) == true
-            || dockBySurface != nil
-            || dockByPane != nil
-        guard routesWindowDock else {
-            return (false, nil, nil)
-        }
-
-        let dockByOwner = requestedWorkspaceID.flatMap { AppDelegate.shared?.existingWindowDock(forWindowId: $0) }
-        // Explicit selectors that name two different windows' Docks fail closed
-        // rather than silently acting on one of them. (The legacy alias pins no
-        // specific Dock, so it never conflicts.)
-        if windowDockSelectorsConflict(
-            requestedWorkspaceID: requestedWorkspaceID,
-            requestedWindowID: v2UUID(params, "window_id"),
-            dockByOwner: dockByOwner,
-            dockBySurface: dockBySurface,
-            dockByPane: dockByPane
-        ) {
-            return (true, nil, .err(code: "invalid_params", message: dockConflictingRoutingSelectorsMessage(), data: nil))
-        }
-        let dock = dockBySurface
-            ?? dockByPane
-            ?? dockByOwner
-            ?? AppDelegate.shared?.windowDock(for: tabManager)
-        guard let dock else {
-            return (true, nil, .err(code: "not_found", message: "Workspace not found", data: nil))
-        }
-        return (true, dock, nil)
-    }
-
-    /// Whether explicit Dock selectors name more than one distinct window's
-    /// Dock: owner `workspace_id` vs surface vs pane vs an explicit `window_id`
-    /// (a window Dock's owner id IS its window id). A `workspace_id` that names
-    /// a NON-Dock scope (neither the legacy alias nor a Dock owner) likewise
-    /// contradicts a Dock resolved from surface/pane selectors. Browser CLI
-    /// commands never inject the caller's workspace context (unlike e.g.
-    /// `close-surface`), so this stays fail-closed without breaking the CLI.
-    private func windowDockSelectorsConflict(
-        requestedWorkspaceID: UUID?,
-        requestedWindowID: UUID?,
-        dockByOwner: DockSplitStore?,
-        dockBySurface: DockSplitStore?,
-        dockByPane: DockSplitStore?
-    ) -> Bool {
-        let resolved = [dockByOwner, dockBySurface, dockByPane].compactMap { $0 }
-        guard let first = resolved.first else { return false }
-        if resolved.contains(where: { $0 !== first }) { return true }
-        if let requestedWindowID, first.workspaceId != requestedWindowID { return true }
-        if let requestedWorkspaceID,
-           requestedWorkspaceID != AppDelegate.windowDockAliasWorkspaceId,
-           dockByOwner == nil {
-            return true
-        }
-        return false
-    }
-
     private func v2BrowserTabListPayload(
         workspaceId: UUID,
         focusedPanelId: UUID?,
@@ -5467,16 +5326,6 @@ class TerminalController {
             "surface_ref": v2Ref(kind: .surface, uuid: focusedPanelId),
             "tabs": tabs
         ]
-    }
-
-    private func closeWindowDockBrowserPanel(_ targetId: UUID, in dock: DockSplitStore) -> Bool {
-        guard let tabId = dock.surfaceId(forPanelId: targetId) else { return false }
-        dock.forceCloseDockTabIds.insert(tabId)
-        let closed = dock.bonsplitController.closeTab(tabId)
-        if !closed {
-            dock.forceCloseDockTabIds.remove(tabId)
-        }
-        return closed
     }
 
     private nonisolated func v2BrowserPanelFields(_ ctx: V2BrowserPanelContext, adding fields: [String: Any] = [:]) -> [String: Any] {
@@ -7844,26 +7693,6 @@ class TerminalController {
         )
     }
 
-    /// Dock-scoped browser action payload. A window Dock's owner id IS its
-    /// window id, so the reported window comes from the Dock itself rather than
-    /// the routed `tabManager`, which can resolve a different window when the
-    /// caller's injected context disagrees with the surface's home.
-    private func v2WindowDockBrowserActionPayload(
-        _ ctx: V2BrowserPanelContext,
-        extra: [String: Any] = [:]
-    ) -> [String: Any] {
-        var payload: [String: Any] = [
-            "workspace_id": ctx.workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: ctx.workspaceId),
-            "surface_id": ctx.surfaceId.uuidString,
-            "surface_ref": v2Ref(kind: .surface, uuid: ctx.surfaceId),
-            "window_id": ctx.workspaceId.uuidString,
-            "window_ref": v2Ref(kind: .window, uuid: ctx.workspaceId)
-        ]
-        for (key, value) in extra { payload[key] = value }
-        return payload
-    }
-
     /// Returns an error if any of the given handle params is SUPPLIED but does not resolve.
     /// v2UUID returns nil for both an absent param and a present-but-unresolvable handle (e.g. a
     /// stale `surface:2`/`workspace:99` ref), so a supplied target must not be treated as omitted
@@ -8133,13 +7962,7 @@ class TerminalController {
             let browserPanel = context.browserPanel
 
             if let windowDock = windowDockContainingPanel(surfaceId) {
-                // A Dock surface renders only in its owning window (the Dock
-                // registry is the source of truth), so focus and reveal there
-                // even if the caller's routed context named another window.
-                let owningTabManager = dockOwnerTabManager(for: windowDock, fallback: tabManager)
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowDock.workspaceId)
-                setActiveTabManager(owningTabManager)
-                revealDockForFocus(tabManager: owningTabManager)
+                _ = focusAndRevealWindowDock(for: windowDock, fallback: tabManager)
                 windowDock.focusPanel(surfaceId)
             } else {
                 if let windowId = v2ResolveWindowId(tabManager: tabManager) {
