+++ cmuxTests/DockSocketLifecycleTests.swift
@@ -418,41 +418,6 @@ struct DockSocketLifecycleTests {
         }
     }
 
-    @Test("Legacy global Dock alias workspace_id routes to the caller window's Dock")
-    @MainActor
-    func legacyDockAliasRoutesToCallerWindowDock() throws {
-        try withDockEnabled {
-            try withSocketAppContext { _, _, windowId in
-                let createResult = try v2Result(
-                    method: "surface.create",
-                    params: ["placement": "dock", "type": "terminal", "focus": true]
-                )
-                let dockSurfaceIdRaw = try #require(createResult["dock_surface_id"] as? String)
-                let dockSurfaceId = try #require(UUID(uuidString: dockSurfaceIdRaw))
-
-                // The retired app-wide Global Dock's owner id keeps routing for
-                // CLI compatibility: it resolves to the Dock of the window the
-                // command targets, and results report that Dock's real owner id.
-                let listResult = try v2Result(
-                    method: "surface.list",
-                    params: ["workspace_id": AppDelegate.windowDockAliasWorkspaceId.uuidString]
-                )
-                let surfaces = try #require(listResult["surfaces"] as? [[String: Any]])
-                #expect(listResult["workspace_id"] as? String == windowId.uuidString)
-                #expect(surfaces.contains { $0["id"] as? String == dockSurfaceId.uuidString })
-
-                let closeResult = try v2Result(
-                    method: "surface.close",
-                    params: ["workspace_id": AppDelegate.windowDockAliasWorkspaceId.uuidString]
-                )
-                #expect(closeResult["workspace_id"] as? String == windowId.uuidString)
-                #expect(closeResult["surface_id"] as? String == dockSurfaceId.uuidString)
-                let windowDock = try #require(AppDelegate.shared?.existingWindowDock(forWindowId: windowId))
-                #expect(!windowDock.containsPanel(dockSurfaceId))
-            }
-        }
-    }
-
     @Test("Window Dock owner pane mutations do not fall back to selected workspace")
     @MainActor
     func windowDockOwnerPaneMutationDoesNotFallBackToSelectedWorkspace() throws {
@@ -685,152 +650,6 @@ struct DockSocketLifecycleTests {
             #expect(!store.containsPanel(panelId))
             #expect(!notificationStore.hasUnreadNotification(forTabId: workspace.id, surfaceId: panelId))
         }
-    }
-
-    @Test("Window Dock pane routing and focused close stay in their own window's Dock")
-    @MainActor
-    func dockPaneRoutingAndFocusedCloseStayInDock() throws {
-#if DEBUG
-        try withDockEnabled {
-            let previousAppDelegate = AppDelegate.shared
-            let previousManager = TerminalController.shared.activeTabManagerForCallerNotification()
-            let appDelegate = AppDelegate()
-            let activeManager = TabManager(autoWelcomeIfNeeded: false)
-            let dockManager = TabManager(autoWelcomeIfNeeded: false)
-            let fileExplorerState = FileExplorerState()
-            let activeWindowId = UUID()
-            let activeWindow = NSWindow(contentRect: NSRect(x: 0, y: 0, width: 640, height: 480), styleMask: [.titled, .closable], backing: .buffered, defer: false)
-            activeWindow.isReleasedWhenClosed = false
-            activeWindow.identifier = NSUserInterfaceItemIdentifier("cmux.main.\(activeWindowId.uuidString)")
-            let dockWindowId = UUID()
-            let dockWindow = NSWindow(contentRect: NSRect(x: 0, y: 0, width: 320, height: 240), styleMask: [.titled, .closable], backing: .buffered, defer: false)
-            dockWindow.isReleasedWhenClosed = false
-            dockWindow.identifier = NSUserInterfaceItemIdentifier("cmux.main.\(dockWindowId.uuidString)")
-
-            AppDelegate.shared = appDelegate
-            appDelegate.tabManager = activeManager
-            TerminalController.shared.setActiveTabManager(activeManager)
-            appDelegate.registerMainWindow(activeWindow, windowId: activeWindowId, tabManager: activeManager, sidebarState: SidebarState(), sidebarSelectionState: SidebarSelectionState())
-            appDelegate.registerMainWindow(dockWindow, windowId: dockWindowId, tabManager: dockManager, sidebarState: SidebarState(), sidebarSelectionState: SidebarSelectionState(), fileExplorerState: fileExplorerState)
-            dockWindow.orderFront(nil)
-            defer {
-                TerminalController.shared.setActiveTabManager(previousManager)
-                // Unregistering each window context also tears down its Dock.
-                appDelegate.unregisterMainWindowContextForTesting(windowId: activeWindowId)
-                appDelegate.unregisterMainWindowContextForTesting(windowId: dockWindowId)
-                activeManager.tabs.forEach { $0.teardownAllPanels() }
-                dockManager.tabs.forEach { $0.teardownAllPanels() }
-                activeWindow.orderOut(nil)
-                activeWindow.close()
-                dockWindow.orderOut(nil)
-                dockWindow.close()
-                AppDelegate.shared = previousAppDelegate
-            }
-
-            let activeWorkspace = try #require(activeManager.tabs.first)
-            let dockWorkspace = try #require(dockManager.tabs.first)
-            let activeWindowDock = appDelegate.windowDock(forWindowId: activeWindowId)
-            let dockPane = try #require(activeWindowDock.bonsplitController.allPaneIds.first)
-            let result = try v2Result(method: "surface.create", params: [
-                "placement": "dock",
-                "type": "terminal",
-                "pane_id": dockPane.id.uuidString,
-                "window_id": activeWindowId.uuidString,
-            ])
-            let dockSurfaceIdRaw = try #require(result["dock_surface_id"] as? String)
-            let dockSurfaceId = try #require(UUID(uuidString: dockSurfaceIdRaw))
-            #expect(result["window_id"] as? String == activeWindowId.uuidString)
-            #expect(result["workspace_id"] as? String == activeWindowId.uuidString)
-            #expect(activeWindowDock.containsPanel(dockSurfaceId))
-            #expect(appDelegate.existingWindowDock(forWindowId: dockWindowId) == nil)
-            #expect(activeWorkspace._dockSplit?.containsPanel(dockSurfaceId) != true)
-            #expect(dockWorkspace._dockSplit?.containsPanel(dockSurfaceId) != true)
-
-            // Seed the second window's own Dock so the focused close below has a
-            // panel to act on in ITS window.
-            let otherWindowDock = appDelegate.windowDock(forWindowId: dockWindowId)
-            #expect(otherWindowDock !== activeWindowDock)
-            let otherPane = try #require(otherWindowDock.resolvePane(requestedPaneID: nil))
-            let otherDockSurfaceId = try #require(otherWindowDock.newSurface(kind: .terminal, inPane: otherPane, focus: true))
-
-            // An explicit window_id that contradicts the Dock named by
-            // workspace_id fails closed instead of re-homing the report.
-            let crossListEnvelope = try v2Envelope(method: "surface.list", params: [
-                "workspace_id": activeWindowId.uuidString,
-                "window_id": dockWindowId.uuidString,
-            ])
-            #expect(crossListEnvelope["ok"] as? Bool == false)
-
-            // A Dock-owner workspace_id alone routes to (and reports) the
-            // owning window, wherever the caller runs.
-            let ownerList = try v2Result(method: "surface.list", params: [
-                "workspace_id": activeWindowId.uuidString,
-            ])
-            #expect(ownerList["workspace_id"] as? String == activeWindowId.uuidString)
-            #expect(ownerList["window_id"] as? String == activeWindowId.uuidString)
-
-            // Explicitly naming two different windows' Docks fails closed.
-            let conflictEnvelope = try v2Envelope(method: "browser.tab.list", params: [
-                "workspace_id": activeWindowId.uuidString,
-                "surface_id": otherDockSurfaceId.uuidString,
-            ])
-            #expect(conflictEnvelope["ok"] as? Bool == false)
-            let conflictError = try #require(conflictEnvelope["error"] as? [String: Any])
-            #expect(conflictError["message"] as? String == "Conflicting Dock routing selectors")
-
-            // A workspace_id naming a NON-Dock scope contradicts a Dock surface
-            // selector the same way.
-            let workspaceScopeConflict = try v2Envelope(method: "browser.tab.list", params: [
-                "workspace_id": dockWorkspace.id.uuidString,
-                "surface_id": otherDockSurfaceId.uuidString,
-            ])
-            #expect(workspaceScopeConflict["ok"] as? Bool == false)
-
-            // Focusing a Dock surface by id targets its owning window even when
-            // the caller's context resolved elsewhere; an explicit contradictory
-            // window_id fails closed instead.
-            let crossFocusEnvelope = try v2Envelope(method: "surface.focus", params: [
-                "surface_id": dockSurfaceId.uuidString,
-                "window_id": dockWindowId.uuidString,
-            ])
-            #expect(crossFocusEnvelope["ok"] as? Bool == false)
-            let focusResult = try v2Result(method: "surface.focus", params: [
-                "surface_id": dockSurfaceId.uuidString,
-            ])
-            #expect(focusResult["window_id"] as? String == activeWindowId.uuidString)
-            #expect(focusResult["workspace_id"] as? String == activeWindowId.uuidString)
-            #expect(activeWindowDock.focusedPanelId == dockSurfaceId)
-
-            let closeAction = KeyboardShortcutSettings.Action.closeTab
-            let hadCloseShortcut = UserDefaults.standard.object(forKey: closeAction.defaultsKey) != nil
-            let originalCloseShortcut = KeyboardShortcutSettings.shortcut(for: closeAction)
-            KeyboardShortcutSettings.setShortcut(closeAction.defaultShortcut, for: closeAction)
-            defer {
-                if hadCloseShortcut {
-                    KeyboardShortcutSettings.setShortcut(originalCloseShortcut, for: closeAction)
-                } else {
-                    KeyboardShortcutSettings.resetShortcut(for: closeAction)
-                }
-            }
-
-            let mainPanelCount = dockWorkspace.panels.count
-            let focusedMainPanel = dockWorkspace.focusedPanelId
-            fileExplorerState.setVisible(true)
-            fileExplorerState.mode = .dock
-            appDelegate.noteRightSidebarKeyboardFocusIntent(mode: .dock, in: dockWindow)
-            let closeEvent = try #require(NSEvent.keyEvent(with: .keyDown, location: .zero, modifierFlags: [.command], timestamp: ProcessInfo.processInfo.systemUptime, windowNumber: dockWindow.windowNumber, context: nil, characters: "w", charactersIgnoringModifiers: "w", isARepeat: false, keyCode: 13))
-
-            #expect(appDelegate.debugHandleCustomShortcut(event: closeEvent))
-            // The focused close acted on the second window's OWN Dock; the first
-            // window's Dock panel is untouched.
-            #expect(!otherWindowDock.containsPanel(otherDockSurfaceId))
-            #expect(activeWindowDock.containsPanel(dockSurfaceId))
-            #expect(dockWorkspace.panels.count == mainPanelCount)
-            #expect(dockWorkspace.focusedPanelId == focusedMainPanel)
-        }
-#else
-        Issue.record("debugHandleCustomShortcut is only available in DEBUG")
-#endif
     }
 
     @Test("Runtime close routes Dock terminals through the Dock lifecycle")
